Add files via upload
I'm here creating the pipeline model of the program to collect information from the Wikipedia page. Program work is based on the Name and Nicknames of famous personalities worldwide when the user designates the name in the command window.
The current project shows information related to the person in paragraph format. 
# Progamning Language
Python	 
# The primary usage of the library are :
1. Requests Library
2. bs4 
3. Beautiful soup
4. .get() 
5. .split()
6. .remove()
7. .find() and . find_all()
8. .text